pyproject.toml
[tool.poetry]
name = "ArchPackageBot"
version = "0.1.0"
description = "a qqBot for query packages on archlinux"
authors = ["fansuregrin <1134366185@qq.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
httpx = "^0.21.1"
python = "^3.7.3"
nonebot2 = "^2.0.0.a1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.5.0"
nonebot-plugin-test = "^0.3.0"

[nonebot.plugins]
plugins = []
plugin_dirs = ["ArchPackageBot/plugins"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
